renamed package domain -> model
